REFACTOR: Directory 구조 변경
