+++ joonggo-app/src/app/_components/products/product-list.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import {useSearchParams} from "next/navigation";
-import ProductFilter from "@/app/products/[[...category]]/_components/product-filter";
+import ProductFilter from "@/app/_components/products/product-filter";
 import {useEffect, useState} from "react";
-import {Product} from "@/app/types/Product";
-import {SearchParam} from "@/app/types/SearchParam";
-import {search} from "@/app/actions/products";
+import {Product} from "@/types/Product";
+import {SearchParam} from "@/types/SearchParam";
+import {search} from "@/scripts/products";
 
 export default function ProductList(param: { category: string[] }) {
     const {category} = param;
