+++ joonggo-app/src/types/Product.ts
