+++ joonggo-app/src/types/SearchParam.ts
